requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.1
python-dotenv==1.0.0
redis==5.0.1
aioredis==2.0.1  # For async Redis support
python-json-logger==2.0.7

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0
asyncio==3.4.3
async-timeout==4.0.3 